% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_similarity_mat}
\alias{cosine_similarity_mat}
\title{Cosine Similarities}
\usage{
cosine_similarity_mat(X)
}
\arguments{
\item{X}{sp_mat}
}
\description{
Given a User-Item Sparse Matrix, it will calculate a User-User Sparse Matrix of Cosine Similarities
}
