% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DtFunctions.R
\name{DayToWeek}
\alias{DayToWeek}
\title{Converst to a weekly resolution from a daily one on a financial data table}
\usage{
DayToWeek(dt)
}
\arguments{
\item{dt}{daily resolution data.table}
}
\description{
Converst to a weekly resolution from a daily one on a financial data table
}
