% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coinmarketcap_API.R
\name{get_crypto_dim_map}
\alias{get_crypto_dim_map}
\title{Retrieve all Cryptocurrencies listed on coinmarketcap}
\usage{
get_crypto_dim_map()
}
\description{
Retrieve all Cryptocurrencies listed on coinmarketcap
}
\seealso{
Other CMC API: 
\code{\link{get_crypto_listings_latest}()},
\code{\link{get_crypto_quotes_latest}()}
}
\concept{CMC API}
