% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coinmarketcap_API.R
\name{get_crypto_listings_latest}
\alias{get_crypto_listings_latest}
\title{Retrieve top 100 Cryptocurrencies latest listing}
\usage{
get_crypto_listings_latest()
}
\description{
Retrieve top 100 Cryptocurrencies latest listing
}
\seealso{
Other CMC API: 
\code{\link{get_crypto_dim_map}()},
\code{\link{get_crypto_quotes_latest}()}
}
\concept{CMC API}
