% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read7z}
\alias{read7z}
\title{Use data.table::fread to read a zipped file}
\usage{
read7z(file, seven_zip = sevenzip_path)
}
\arguments{
\item{file}{file to read}

\item{seven_zip}{7z installation path}
}
\description{
Use data.table::fread to read a zipped file
}
\seealso{
Other read_write: 
\code{\link{read7zfeather}()},
\code{\link{write7zfeather}()},
\code{\link{write7z}()}
}
\concept{read_write}
