% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read7zfeather}
\alias{read7zfeather}
\title{Use feather::read_feather to read a zipped file}
\usage{
read7zfeather(file, seven_zip = "C:/Users/user/scoop/apps/7zip/current/7z.exe")
}
\arguments{
\item{file}{file to read}

\item{seven_zip}{7z installation path}
}
\description{
Use feather::read_feather to read a zipped file
}
\seealso{
Other read_write: 
\code{\link{read7z}()},
\code{\link{write7zfeather}()},
\code{\link{write7z}()}
}
\concept{read_write}
