% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simple_moving_avg}
\alias{simple_moving_avg}
\title{Simple moving average}
\usage{
simple_moving_avg(dt, n)
}
\arguments{
\item{dt}{data table/frame}

\item{n}{lenght of the average}
}
\description{
Given a data table/frame with a column 'x', returns the simple moving average of that column
}
