% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sma}
\alias{sma}
\title{Vector simple moving average}
\usage{
sma(x, n)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{lenght of the average}
}
\description{
Returns a vector of the simple moving average of a vector
}
