% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_moving_avg}
\alias{weighted_moving_avg}
\title{Weighted moving average}
\usage{
weighted_moving_avg(dt, n)
}
\arguments{
\item{dt}{data table/frame}

\item{n}{lenght of the average}
}
\description{
Given a data table/frame with a column 'x', returns the weighted moving average of that column
}
