% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write7z}
\alias{write7z}
\title{Use data.table::fwrite to write and zipped a file}
\usage{
write7z(
  data,
  filename,
  seven_zip = "C:/Users/user/scoop/apps/7zip/current/7z.exe",
  sep = ","
)
}
\arguments{
\item{data}{file to export}

\item{filename}{path to be exported to}

\item{seven_zip}{7z installation path}

\item{sep}{column seperator}
}
\description{
Use data.table::fwrite to write and zipped a file
}
\seealso{
Other read_write: 
\code{\link{read7zfeather}()},
\code{\link{read7z}()},
\code{\link{write7zfeather}()}
}
\concept{read_write}
